+++ src/vlr/BufferViewSet.cpp
@@ -12,7 +12,7 @@ namespace vlr {
             .descriptorCount = uint32_t(buffers.size()),
             .descriptorType = VK_DESCRIPTOR_TYPE_STORAGE_BUFFER
         });
-        for (uintptr_t i = 0; i < buffers.size(); i++) {
+        for (uint32_t i = 0; i < buffers.size(); i++) {
             handle.offset<VkDescriptorBufferInfo>(i) = this->buffers[i];
         };
         vkh::handleVk(vkt::AllocateDescriptorSetWithUpdate(driver->getDeviceDispatch(), this->descriptorSetInfo, this->set, this->updated));
